tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-DOMParser-parseFromString.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>

<meta http-equiv="Content-Security-Policy" content="require-trusted-types">
<body>
<script>
  // Trusted HTML assignments do not throw.
  test(t => {
    let p = createHTML_policy(window, 1);
    let html = p.createHTML(INPUTS.HTML);
    let parser = new DOMParser();
    let doc = parser.parseFromString(html, "text/html");
    assert_equals(doc.body.innerText, RESULTS.HTML);
  }, "document.innerText assigned via policy (successful HTML transformation).");

  // String assignments throw.
  test(t => {
    var parser = new DOMParser();
    assert_throws(new TypeError(), _ => {
      var doc = parser.parseFromString("Fail", "text/html");
    });
  }, "`document.innerText = string` throws.");

  // Null assignment throws.
  test(t => {
    var parser = new DOMParser();
    assert_throws(new TypeError(), _ => {
      var doc = parser.parseFromString(null, "text/html");
    });
  }, "'document.innerText = null' throws");
</script>
